Removed debug logging

+++ npstructures/raggedshape.py
@@ -3,7 +3,7 @@ import numpy as np
 import logging
 
 class ViewBase:
-    _dtype = np.int32
+    _dtype = np.int64
 
     @classmethod
     def set_dtype(cls, dtype):
